Extract JSON filtering functionality from `gh api`
